\name{gfm_table}
\alias{gfm_table}
\title{Print a table in github-flavored markdown}
\usage{
  gfm_table(x, ...)
}
\arguments{
  \item{x}{an R table object}

  \item{...}{aditional arguments to ascii}
}
\value{
  prints a gfm marked up table (nice ascii readable format)
}
\description{
  Print a table in github-flavored markdown
}

