\name{render_wordpress}
\alias{render_wordpress}
\title{A function to set the wordpress rendering environment with code syntax highlighting}
\usage{
  render_wordpress(upload = TRUE,
    image_service = c("wordpress", "imgur", "flickr"))
}
\arguments{
  \item{upload}{should images be uploaded}

  \item{image_service}{which image method should be used?
  See details}
}
\value{
  creates a wordpress rendering environment
}
\description{
  A function to set the wordpress rendering environment
  with code syntax highlighting
}
\details{
  wordpress
}

